Clean up type usage

+++ rw/src/load_texture.c
@@ -1,16 +1,16 @@
-#include <WINDOWS.H>
+#include "types.h"
 
-void bmp_pallete_rgb888_to_rgb1555(byte *pallete) {
+void bmp_pallete_rgb888_to_rgb1555(uint8 *pallete) {
     int i;
-    byte r, g, b;
-    byte *ptr = pallete + 2;
+    uint8 r, g, b;
+    uint8 *ptr = pallete + 2;
 
     for (i = 0; i < 256; i++) {
-        r = ptr[-2] / 8;
-        g = ptr[-1] / 8;
-        b = ptr[0]  / 8;
+        r = (uint8)(ptr[-2] / 8);
+        g = (uint8)(ptr[-1] / 8);
+        b = (uint8)(ptr[0]  / 8);
 
-        *((short*)pallete) = (r << 5 | g) << 5 | b;
+        *((uint16*)pallete) = (uint16)((r << 5 | g) << 5 | b);
 
         ptr += 3;
         pallete += 2;
